# server/src/main/java/at/bestsolution/quti/service/model/DateTimeRange.java
// Generated by RSD - Do not modify
package at.bestsolution.quti.service.model;

import java.time.ZonedDateTime;

public interface DateTimeRange {

	public interface Data extends _Base.BaseData, DateTimeRange {
		public ZonedDateTime min();

		public ZonedDateTime max();

	}

	public interface DataBuilder extends _Base.BaseDataBuilder<DateTimeRange.Data> {
		public DataBuilder min(ZonedDateTime min);

		public DataBuilder max(ZonedDateTime max);

	}
}