# server/src/main/java/at/bestsolution/quti/service/model/EventSearch.java
// Generated by RSD - Do not modify
package at.bestsolution.quti.service.model;

import java.util.function.Function;
import java.util.List;

public interface EventSearch {

	public interface Data extends _Base.BaseData, EventSearch {
		public List<String> tags();

		public DateTimeRange.Data startRange();

		public DateTimeRange.Data endRange();

	}

	public interface DataBuilder extends _Base.BaseDataBuilder<EventSearch.Data> {
		public DataBuilder tags(List<String> tags);

		public DataBuilder startRange(DateTimeRange.Data startRange);

		public <T extends DateTimeRange.DataBuilder> DataBuilder withStartRange(Class<T> clazz, Function<T, DateTimeRange.Data> block);

		public DataBuilder endRange(DateTimeRange.Data endRange);

		public <T extends DateTimeRange.DataBuilder> DataBuilder withEndRange(Class<T> clazz, Function<T, DateTimeRange.Data> block);

	}
}