# server/src/main/resources/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Quti",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/calendar/{calendar}/events": {
      "post": {
        "tags": [
          "Event"
        ],
        "description": " create a new event",
        "parameters": [
          {
            "name": "calendar",
            "description": "key of the calendar owning the new event",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventNew"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "key of the new event",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/calendar/{calendar}/events/{key}": {
      "get": {
        "tags": [
          "Event"
        ],
        "description": " get an event",
        "parameters": [
          {
            "name": "calendar",
            "description": "key of the calendar owning the event",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key",
            "description": "key of the event",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "timezone",
            "description": "the timezone",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "the event",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Event"
        ],
        "description": " update an event",
        "parameters": [
          {
            "name": "calendar",
            "description": "key of the calendar owning the event",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key",
            "description": "key of the event",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventPatch"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "success"
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Event"
        ],
        "description": " delete an event",
        "parameters": [
          {
            "name": "calendar",
            "description": "key of the calendar owning the event",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key",
            "description": "key of the event",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "success"
          }
        }
      }
    },
    "/api/calendar/{calendar}/events/search": {
      "put": {
        "tags": [
          "Event"
        ],
        "description": "",
        "parameters": [
          {
            "name": "calendar",
            "description": "key of the calendar owning the event",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "timezone",
            "description": "the timezone the date",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventSearch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Event"
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/calendar/{calendar}/events/{key}/action/cancel": {
      "put": {
        "tags": [
          "Event"
        ],
        "description": " cancel an event",
        "parameters": [
          {
            "name": "calendar",
            "description": "key of the calendar owning the event",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key",
            "description": "key of the event",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "success"
          }
        }
      }
    },
    "/api/calendar/{calendar}/events/{key}/action/uncancel": {
      "put": {
        "tags": [
          "Event"
        ],
        "description": " uncancel an event",
        "parameters": [
          {
            "name": "calendar",
            "description": "key of the calendar owning the event",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key",
            "description": "key of the event",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "success"
          }
        }
      }
    },
    "/api/calendar/{calendar}/events/{key}/action/move": {
      "put": {
        "tags": [
          "Event"
        ],
        "description": " move an event to a new time",
        "parameters": [
          {
            "name": "calendar",
            "description": "key of the calendar owning the event",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key",
            "description": "key of the event",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "start": {
                    "type": "string",
                    "format": "zoned-date-time"
                  },
                  "end": {
                    "type": "string",
                    "format": "zoned-date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "success"
          }
        }
      }
    },
    "/api/calendar/{calendar}/events/{key}/action/end-repeat": {
      "put": {
        "tags": [
          "Event"
        ],
        "description": " end a repeating event",
        "parameters": [
          {
            "name": "calendar",
            "description": "key of the calendar owning the event",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key",
            "description": "key of the event",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "pattern": "\\d\\d\\d\\d-\\d\\d-\\d\\d",
                "format": "date"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "success"
          }
        }
      }
    },
    "/api/calendar/{calendar}/events/{key}/action/description": {
      "put": {
        "tags": [
          "Event"
        ],
        "description": " set a description",
        "parameters": [
          {
            "name": "calendar",
            "description": "key of the calendar owning the event",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key",
            "description": "key of the event",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DayOfWeek": {
        "type": "string",
        "enum": [
          "MONDAY",
          "TUESDAY",
          "WEDNESDAY",
          "THURSDAY",
          "FRIDAY",
          "SATURDAY",
          "SUNDAY"
        ]
      },
      "Month": {
        "type": "string",
        "enum": [
          "JANUARY",
          "FEBRUARY",
          "MARCH",
          "APRIL",
          "MAY",
          "JUNE",
          "JULY",
          "AUGUST",
          "SEPTEMBER",
          "OCTOBER",
          "DECEMBER"
        ]
      },
...
      "EventRepeatDaily": {
        "type": "object",
        "properties": {
          "interval": {
            "type": "number",
            "format": "int32"
          },
          "endDate": {
            "type": "string",
            "pattern": "\\d\\d\\d\\d-\\d\\d-\\d\\d",
            "format": "date"
          },
          "timeZone": {
            "type": "string"
          },
          "@type": {
            "type": "string"
          }
        },
        "required": [
          "endDate"
        ]
      },
...
      "EventRepeatWeekly": {
        "type": "object",
        "properties": {
          "daysOfWeek": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DayOfWeek"
            }
          },
          "interval": {
            "type": "number",
            "format": "int32"
          },
          "endDate": {
            "type": "string",
            "pattern": "\\d\\d\\d\\d-\\d\\d-\\d\\d",
            "format": "date"
          },
          "timeZone": {
            "type": "string"
          },
          "@type": {
            "type": "string"
          }
        },
        "required": [
          "endDate"
        ]
      },
...
      "EventRepeatAbsoluteMonthly": {
        "type": "object",
        "properties": {
          "dayOfMonth": {
            "type": "number",
            "format": "int32"
          },
          "interval": {
            "type": "number",
            "format": "int32"
          },
          "endDate": {
            "type": "string",
            "pattern": "\\d\\d\\d\\d-\\d\\d-\\d\\d",
            "format": "date"
          },
          "timeZone": {
            "type": "string"
          },
          "@type": {
            "type": "string"
          }
        },
        "required": [
          "endDate"
        ]
      },
...
      "EventRepeatAbsoluteYearly": {
        "type": "object",
        "properties": {
          "dayOfMonth": {
            "type": "number",
            "format": "int32"
          },
          "month": {
            "$ref": "#/components/schemas/Month"
          },
          "interval": {
            "type": "number",
            "format": "int32"
          },
          "endDate": {
            "type": "string",
            "pattern": "\\d\\d\\d\\d-\\d\\d-\\d\\d",
            "format": "date"
          },
          "timeZone": {
            "type": "string"
          },
          "@type": {
            "type": "string"
          }
        },
        "required": [
          "endDate"
        ]
      },
...
      "EventRepeatRelativeMonthly": {
        "type": "object",
        "properties": {
          "daysOfWeek": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DayOfWeek"
            }
          },
          "interval": {
            "type": "number",
            "format": "int32"
          },
          "endDate": {
            "type": "string",
            "pattern": "\\d\\d\\d\\d-\\d\\d-\\d\\d",
            "format": "date"
          },
          "timeZone": {
            "type": "string"
          },
          "@type": {
            "type": "string"
          }
        },
        "required": [
          "endDate"
        ]
      },
...
      "EventRepeatRelativeYearly": {
        "type": "object",
        "properties": {
          "daysOfWeek": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DayOfWeek"
            }
          },
          "month": {
            "$ref": "#/components/schemas/Month"
          },
          "interval": {
            "type": "number",
            "format": "int32"
          },
          "endDate": {
            "type": "string",
            "pattern": "\\d\\d\\d\\d-\\d\\d-\\d\\d",
            "format": "date"
          },
          "timeZone": {
            "type": "string"
          },
          "@type": {
            "type": "string"
          }
        },
        "required": [
          "endDate"
        ]
      },
...
      "EventRepeat": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/EventRepeatDaily"
          },
          {
            "$ref": "#/components/schemas/EventRepeatWeekly"
          },
          {
            "$ref": "#/components/schemas/EventRepeatAbsoluteMonthly"
          },
          {
            "$ref": "#/components/schemas/EventRepeatAbsoluteYearly"
          },
          {
            "$ref": "#/components/schemas/EventRepeatRelativeMonthly"
          },
          {
            "$ref": "#/components/schemas/EventRepeatRelativeYearly"
          }
        ],
        "discriminator": {
          "propertyName": "@type",
          "mapping": {
            "daily": "#/components/schemas/EventRepeatDaily",
            "weekly": "#/components/schemas/EventRepeatWeekly",
            "absolute-monthly": "#/components/schemas/EventRepeatAbsoluteMonthly",
            "absolute-yearly": "#/components/schemas/EventRepeatAbsoluteYearly",
            "relative-monthly": "#/components/schemas/EventRepeatRelativeMonthly",
            "relative-yearly": "#/components/schemas/EventRepeatRelativeYearly"
          }
        }
      },
      "EventNew": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "start": {
            "type": "string",
            "format": "zoned-date-time"
          },
          "end": {
            "type": "string",
            "format": "zoned-date-time"
          },
          "fullday": {
            "type": "boolean"
          },
          "repeat": {
            "$ref": "#/components/schemas/EventRepeat"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "referencedCalendars": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "description",
          "fullday",
          "repeat"
        ]
      },
      "Event": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "start": {
            "type": "string",
            "format": "zoned-date-time"
          },
          "end": {
            "type": "string",
            "format": "zoned-date-time"
          },
          "fullday": {
            "type": "boolean"
          },
          "repeat": {
            "$ref": "#/components/schemas/EventRepeat"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "referencedCalendars": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "key",
          "description",
          "fullday",
          "repeat"
        ]
      },
      "EventPatch": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "start": {
            "type": "string",
            "format": "zoned-date-time"
          },
          "end": {
            "type": "string",
            "format": "zoned-date-time"
          },
          "fullday": {
            "type": "boolean",
            "nullable": true
          },
          "repeat": {
            "$ref": "#/components/schemas/EventRepeat",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "referencedCalendars": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "EventSearch": {
        "type": "object",
        "properties": {
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "startRange": {
            "$ref": "#/components/schemas/DateTimeRange"
          },
          "endRange": {
            "$ref": "#/components/schemas/DateTimeRange"
          }
        },
        "required": [
          "tags",
          "startRange",
          "endRange"
        ]
      },
      "DateTimeRange": {
        "type": "object",
        "properties": {
          "min": {
            "type": "string",
            "format": "zoned-date-time"
          },
          "max": {
            "type": "string",
            "format": "zoned-date-time"
          }
        },
        "required": [
          "min",
          "max"
        ]
      }
    }
  }
}